requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
clearbit==0.1.7
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
idna==3.3
psycopg2==2.9.3
PyJWT==1.7.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
